Caesar Cipher Initial Upload
Basic Caesar Cipher with pre-determined string and shift. Want to add user input for string and shift in future. Initially thought of making a nested array with an array with only capital alphabet and another with lowercase and having it pushed (shifted) based on shift. Might try that out in a test case later if seems viable.